Document options_command and fold launcher tests into their source files
The quick-actions README covered [[options]] but not options_command, so
the dynamic-list form and its tab-separated description syntax were
undocumented.

launch_test.go held tests for launchProjects and launchQuickActions, which
live in projects.go and quickactions.go; this repo keeps one test file per
source file.
